% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P2C2M_Skyline.R
\name{P2C2M.Skyline}
\alias{P2C2M.Skyline}
\title{Run P2C2M.Skyline}
\usage{
P2C2M.Skyline(tree.file, log.file, dir, nrep, path.to.ms)
}
\arguments{
\item{tree.file}{A ultrametric phylogenetic tree (NEXUS format).}

\item{log.file}{The log file resulting from a Bayesian Skyline analyzed in Tracer.}

\item{dir}{Set the working directory.}

\item{nrep}{The size of the null distribution, Default: 100.}

\item{path.to.ms}{Path to ms software.}
}
\value{
A p-value that specifies whether the empirical dataset is or not a poor fit to the Skyline model.
}
\description{
Assessing model adequacy for Bayesian Skyline Plots using posterior predictive simulation.
}
\author{
Emanuel M. Fonseca, Drew J. Duckett, Filipe G. Almeida, Megan L. Smith, Maria Tereza C. Thomé, Bryan C. Carstens
}
